you can now see without subtitles if you did not put anything

+++ bashflix.sh
@@ -12,23 +12,29 @@ pirate-get -s SeedersDsc -0 -M $1
 magnet=$(find . -maxdepth 1 -name "*.magnet" | head -1)
 name=${magnet:2:-7}
 
-# get subtitles
-get-subs $lang1
-sub=$(find . -maxdepth 1 -name "*.srt" | head -1)
+if [ -z "$lang1" ]; then
+  # get magnet to stream
+  pirate-get -s SeedersDsc -0 -C "peerflix \"%s\" -k" $1
 
-# try to search in english if portuguese not found
-if [ -z "$sub" ]; then
-  get-subs $lang2
+else
+  # get subtitles
+  get-subs $lang1
   sub=$(find . -maxdepth 1 -name "*.srt" | head -1)
+
+  # try to search in english if chosen language not found
+  if [ -z "$sub" ]; then
+    get-subs $lang2
+    sub=$(find . -maxdepth 1 -name "*.srt" | head -1)
+  fi
+
+  # remove parenthesis from subtitles filename
+  sub2=$(echo "$sub" | sed -re 's/[()]//g')
+  sub3=${sub2:2}
+  mv ${sub:2} ${sub3}
+
+  # get magnet to stream
+  pirate-get -s SeedersDsc -0 -C "peerflix \"%s\" -k -t ${sub3}" $1
+
+  # remove created files
+  rm *.srt && rm *.magnet
 fi
-
-# remove parenthesis from subtitles filename
-sub2=$(echo "$sub" | sed -re 's/[()]//g')
-sub3=${sub2:2}
-mv ${sub:2} ${sub3}
-
-# get magnet to stream
-pirate-get -s SeedersDsc -0 -C "peerflix \"%s\" -k -t ${sub3}" $1
-
-# remove created files
-rm *.srt && rm *.magnet
